Refactor AlterReadwriteSplittingRuleStatementUpdaterTest
